Reverting the backports made in r28359 from 2017-05-10. The changes are too comprehensive for the old stable branch and would have increased the minimum required libgcrypt version, causing builds to fail in previously supported platforms.
* nasl/nasl.c, nasl/nasl_cert.c, nasl/nasl_crypto.c,
nasl/nasl_crypto.h, nasl/nasl_crypto2.c, nasl/nasl_crypto2.h,
nasl/nasl_init.c, nasl/nasl_text_utils.c: Revert changes made in
r28359.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/branches/openvas-libraries-8.0@29550 423fd1db-d629-0410-8442-d21db03e70f4

+++ nasl/nasl_text_utils.c
@@ -1346,7 +1346,7 @@ nasl_str_replace (lex_ctxt * lexic)
 tree_cell *
 nasl_int (lex_ctxt * lexic)
 {
-  long int r = get_int_var_by_num (lexic, 0, 0);
+  int r = get_int_var_by_num (lexic, 0, 0);
   tree_cell *retc;
 
   retc = alloc_tree_cell (0, NULL);
